取消邮箱激活字段，只要设置了邮箱就是激活了

+++ app/Http/Controllers/UserEmailController.php
@@ -1,50 +1,61 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use App\Http\Requests\MailRequest;
 use App\Models\EmailToken;
 use App\Models\User;
-use App\Notifications\ActivateEmail;
-use App\Services\UserMailsService;
 use App\Services\UsersService;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
-use Illuminate\Notifications\Messages\MailMessage;
 
 class UserEmailController extends Controller
 {
     // 发送一封激活邮件
-    public function send(UserMailsService $mailsService) {
-        $user = $this->user();
-        if (!$user->email) {
-            return $this->response->errorBadRequest(__('该用户还未设置邮箱'));
+    public function send(Request $request, UsersService $usersService) {
+
+        $this->validate($request, ['email' => 'required|string|email']);
+
+        if($usersService->isEmailBound($request->email)) {
+            return $this->response->errorBadRequest(__('该邮箱已被绑定'));
         }
-        $mailsService->sendActivationMail($user);
+
+        $user = $this->user();
+
+        // 一个用户只能有一条有效token。先清空历史token
+        EmailToken::where('user_id', $user->id)->delete();
+
+        $emailToken = new EmailToken();
+        $emailToken->user_id = $user->id;
+        $emailToken->email = $request->email;
+
+        $token = bcrypt($request->email . time());
+        $emailToken->token = $token;
+        $emailToken->save();
+
+        \Mail::send('emails.activate', ['token' => $token], function ($message) use ($request) {
+            $message->to($request->email)->subject('【Yogooooo】Activation Email');
+        });
+
         return $this->response->noContent();
     }
 
     // 点击激活邮件，绑定邮箱
-    // 未被激活绑定邮箱可以被多个用户填写，谁先激活就是谁的
     public function activate(Request $request) {
         $emailToken = EmailToken::where('token', $request->token)->first();
         if(!$emailToken || $emailToken->updated_at->addDay() < Carbon::now()) {
-            return '激活链接无效或已过期';
+            return 'The activation link is invalid or expired 激活链接无效或已过期';
         }
 
         // 如果已经被绑定了
         if((new UsersService())->isEmailBound($emailToken->email)) {
-            return __('激活链接无效：该邮箱已被绑定');
+            return __('Invalid activation link: this email has been bound 激活链接无效：该邮箱已被绑定');
         }
 
         // 更新该用户的邮箱
-        User::where('id', $emailToken->user_id)->update([
-            'email' => $emailToken->email,
-            'email_activated' => true
-        ]);
+        User::where('id', $emailToken->user_id)->update(['email' => $emailToken->email]);
 
-        $emailToken->delete(); // 删除使用过的token
+        // 删除该用户的所有token
+        EmailToken::where('user_id', $emailToken->user_id)->delete();
 
-        return '激活成功';
+        return 'Successfully activate. Now you can use your email to log in. 激活成功，现在你可以使用邮箱登录';
     }
 }
